Move expand_decorators to its own file to simplify conflicts
