+++ syntex_syntax/src/lib.rs
@@ -101,6 +101,7 @@ pub mod print {
 pub mod ext {
     pub mod base;
     pub mod build;
+    pub mod decorator;
     pub mod env;
     pub mod expand;
     pub mod mtwt;
